+++ samples/01_quickstart.py
@@ -1,6 +1,5 @@
 from common import make_database, reset_collection, sample_key
 
-
 edb = make_database()
 collection_name = sample_key("quickstart")
 people = edb[collection_name]
@@ -14,6 +13,6 @@ try:
     print("document:", person)
     print("name:", person.name)
     print("role:", person.role)
-    print("dict:", dict(person().to_dict()))
+    print("dict:", person().to_dict())
 finally:
     reset_collection(edb, collection_name)